# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wellcome to the Micopia Web Server Swagger Documentation",
        "title": "Micopia Swagger Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/CustomerProductCRUD/createCustomerProduct": {
            "put": {
                "description": "Create CustomerProduct in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomerProduct"
                ],
                "parameters": [
                    {
                        "description": "The CustomerProduct to create",
                        "name": "CreateCustomerProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CustomerProduct"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CustomerProduct"
                        }
                    }
                }
            }
        },
        "/CustomerProductCRUD/getCustomerProduct": {
            "get": {
                "description": "Get CustomerProducts (cart/wishlist) from the database by customer id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomerProduct"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "The Customer identifier",
                        "name": "customerId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CustomerProduct"
                            }
                        }
                    }
                }
            }
        },
...
        "/ProductCRUD/createProduct": {
            "put": {
                "description": "Create Product in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "description": "The Product to create",
                        "name": "CreateProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        },
        "/ProductCRUD/deleteProduct": {
            "delete": {
                "description": "Delete Product in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "The product identifier",
                        "name": "productId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ProductCRUD/getProduct": {
            "get": {
                "description": "Get Product from the database by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "The product identifier",
                        "name": "productId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        },
...
        "/ProductCRUD/updateProduct": {
            "post": {
                "description": "Update Product in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "description": "The Product to update",
                        "name": "UpdateProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateProductRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "models.CustomerProduct": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
    }
}